requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.2
chromadb>=0.3.21
google-generativeai>=0.3.0
python-multipart>=0.0.5
